# Conflicts:
#	SELMs/Api/Controllers/ApiEquipmentHandoverController.cs
#	SELMs/Api/DTOs/EquipmentHandoverFormDTO.cs
#	SELMs/App_Start/MapperConfig.cs
#	SELMs/Models/BusinessModel/ProjectModel.cs
#	SELMs/Repositories/IEquipmentHandoverFormRepository.cs
#	SELMs/Repositories/Implements/EquipmentHandoverFormRepository.cs
#	SELMs/Services/Implements/ProjectService.cs

+++ SELMs/Repositories/Implements/EquipmentHandoverFormRepository.cs
@@ -14,14 +14,27 @@ namespace SELMs.Repositories.Implements
 		private SELMsContext db = new SELMsContext();
 		public void DeleteApplication(int id)
 		{
-			dynamic application = db.Equipment_Handover_Form.Where(a => a.form_id == id).FirstOrDefault();
+			Equipment_Handover_Form application = db.Equipment_Handover_Form.Where(a => a.form_id == id).FirstOrDefault();
 			if (application != null)
 			{
-				List<Equipment_Handover_Form_Detail> applicationDetails = GetApplicationDetail(id);
-				db.Equipment_Handover_Form_Detail.RemoveRange(applicationDetails);
+				List<Equipment_Handover_Form_Detail> applicationDetails = GetApplicationDetailList(application.form_code);
+				if (applicationDetails != null)
+					db.Equipment_Handover_Form_Detail.RemoveRange(applicationDetails);
 				db.Equipment_Handover_Form.Remove(application);
 			}
-			db.SaveChangesAsync();
+			db.SaveChanges();
+		}
+
+		public dynamic GetApplicationDetailById(int id)
+		{
+			dynamic application = null;
+			application = db.Database.Connection.QuerySingleAsync<dynamic>("Proc_GetDetailHEF", new
+			{
+				form_id = id
+
+			}
+				, commandType: CommandType.StoredProcedure).Result;
+			return application;
 		}
 
 		public dynamic GetApplication(int id)
